backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
openai==1.12.0
qdrant-client==1.7.3
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
google-generativeai==0.4.1
dashscope==1.13.1
cohere==5.5.4
zhipuai==2.1.5.20250825
openrouter==0.1.1
transformers==4.38.2
torch==2.2.1
sentencepiece==0.2.0
sentence-transformers==2.5.1
numpy==1.26.4
argos-translate==1.9.1
pydantic==2.6.4
python-multipart==0.0.9
sqlalchemy==2.0.28
psycopg2-binary==2.9.9